xóa import và export những màn nhỏ
