+++ sakai-vue/src/components/Admin/QuanLySize/ViewSize.vue
@@ -103,10 +103,10 @@ const initFilters = () => {
                         </div>
                     </template>
 
-                    <template v-slot:end>
+                    <!-- <template v-slot:end>
                         <FileUpload mode="basic" accept="image/*" :maxFileSize="1000000" label="Import" chooseLabel="Import" class="mr-2 inline-block" />
                         <Button label="Export" icon="pi pi-upload" class="p-button-help" @click="exportCSV($event)" />
-                    </template>
+                    </template> -->
                 </Toolbar>
                 <DataTable
                     ref="dt"
