+++ sakai-vue/src/components/Admin/VatLieu/ThemVatLieu.vue
@@ -3,9 +3,9 @@ import { useToast } from 'primevue/usetoast';
 import { useForm, useField } from 'vee-validate';
 import { ref } from 'vue';
 import * as yup from 'yup';
-import { useVatLieuService } from '../../../service/Admin/VatLieu/VatLieuService';
+import { VatLieuStore } from '../../../service/Admin/VatLieu/VatLieu.api';
 
-const vatLieuService = useVatLieuService();
+const vatLieuService = VatLieuStore();
 const product = ref({});
 const toast = useToast();
 const submitted = ref(false);
